pyproject.toml
[project]
name = "clinical-ml-pipeline"
version = "1.0.0"
description = "GenAI-Powered Clinical Note Summarization and Differential Diagnosis Generation with RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # FastAPI & Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # Pydantic
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # LangChain for RAG
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-google-genai>=0.0.6",
    
    # LLM & Embeddings
    "google-generativeai>=0.3.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    
    # Supabase & Database
    "supabase>=2.3.0",
    "psycopg2-binary>=2.9.9",
    
    # Document Processing
    "PyPDF2>=3.0.1",
    "easyocr>=1.7.0",
    "Pillow>=10.1.0",
    
    # Dataset Loading
    "datasets>=2.16.0",
    "huggingface-hub>=0.20.0",
    
    # Text Processing
    "tiktoken>=0.5.2",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
